+++ v3/crates/graphql/ir/src/order_by.rs
@@ -233,6 +233,7 @@ pub fn build_ndc_order_by_element<'s>(
                         source_type,
                         source_data_connector: data_connector_link,
                         source_type_mappings: type_mappings,
+                        target_model_name,
                         target_source: &target_source.model,
                         target_type,
                         mappings,